moving ui and app things around
