+++ src/lib.rs
@@ -1,5 +1,7 @@
+pub mod app;
 pub mod entity;
 pub mod games;
 pub mod scene;
 pub mod scene_map;
+pub mod ui;
 pub mod user_input;
